refactor: auth 서비스 인터페이스 types.ts로 이동
