--- src/entities/auth/services/LoginResponse.ts
@@ -1,5 +1,0 @@
-export default interface LoginResponse {
-    idToken: string;
-    accessToken: string;
-    refreshToken: string;
-}
