--- src/entities/auth/services/SignupResponse.ts
@@ -1,3 +1,0 @@
-export default interface SignupResponse {
-    message: string;
-}
