+++ src/entities/auth/services/login/login.usage.ts
@@ -1,4 +1,4 @@
-import { login } from "./login";
+import login from "./login";
 
 (async function () {
     try {
