+++ src/entities/auth/services/types.ts
@@ -1,0 +1,14 @@
+export interface LoginRequest {
+    username: string;
+    password: string;
+}
+
+export interface LoginResponse {
+    idToken: string;
+    accessToken: string;
+    refreshToken: string;
+}
+
+export interface SignupResponse {
+    message: string;
+}
