+++ src/features/auth/api/useLoginFetch.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { login } from "@/entities/auth/services/login/login";
+import login from "@/entities/auth/services/login/login";
 import LoginCredential from "@/entities/auth/models/LoginCredential";
 
 export default function useLoginFetch() {
